+++ src/promptflow-evals/CHANGELOG.md
@@ -5,7 +5,15 @@
 ### Features Added
 - Introduced `JailbreakAdversarialSimulator` for customers who need to do run jailbreak and non jailbreak adversarial simulations at the same time. More info in the README.md in `/promptflow/evals/synthetic/README.md#jailbreak-simulator`
 - Exposed batch evaluation run timeout via `PF_BATCH_TIMEOUT_SEC` environment variable. This variable can be used to set the timeout for the batch evaluation for each evaluator and target separately only, not the entire API call.
-
+- Added support for the following languages in the simulator:
+  - Spanish (`es`)
+  - Italian (`it`)
+  - French (`fr`)
+  - German (`de`)
+  - Simplified Chinese (`zh-cn`)
+  - Portuguese (`pt`)
+  - Japanese (`ja`)
+  - English (`en`)
 ### Bugs Fixed
 - Large simulation was causing a jinja exception, this has been fixed.